Reorganise files into qt & nb subpackages
