+++ geoAssembler/qt/editor/fit_object.ui
