+++ geoAssembler/qt/editor/geometry_editor.ui
