+++ geoAssembler/qt/editor/load_detector.ui
