+++ geoAssembler/qt/objects.py
